probe: fix DMA copy-alignment tail and shell output routing
Two correctness fixes for the shell-over-probe-DMA feature:

1. DMA copy-alignment padding (PROBE_DMA_ALIGN=16)

   probe_task drains dmapb with ALIGN_DOWN(avail, copy_align).  On
   ACE15/MTPM copy_align=16, so bytes that would leave avail % 16 != 0
   were permanently un-drainable.  The last 1-15 bytes of every probe
   packet were stuck until the next log burst piggy-backed them.

   Fix: append 0-15 zero bytes after every packet footer in
   probe_write_payload() so that avail is always a multiple of 16.
   The host parser is unaffected because it scans for the magic
   SOF_PROBE_MAGIC header and ignores trailing zeros.

   Also pass dmapb explicitly to probe_gen_header() so the helper
   is no longer hard-coded to ext_dma.

2. Shell output routing (probe_shell_output -> ext_dma)

   Shell packets were being written to shell_dma, but the kernel-side
   compress reader only drains ext_dma (stream_tag 1).  Route shell
   output through ext_dma using PROBE_SHELL_BUFFER_ID (0x02000000) to
   distinguish it from log packets (0x01000000).

   Remove the dead kick_shell_task() helper — shell output now kicks
   the probe task directly via kick_probe_task().

   probe_shell_output() return type changed from ssize_t to int to
   match Zephyr shell backend expectations.

Tested: 'kernel thread list' and 'device list' output fully received
across RTD3 reconnect cycles on ARL-S (ACE15/MTPM).

+++ src/include/sof/probe/probe.h
@@ -42,7 +42,7 @@ void probe_logging_init(probe_logging_hook_t hook);
  * @return Number of bytes written (can be less than length when buffer is
  *         near full), 0 when probes are inactive, negative errno on error.
  */
-ssize_t probe_shell_output(const uint8_t *buffer, size_t length);
+int probe_shell_output(const uint8_t *buffer, size_t length);
 
 /*
  * \brief Initialize probes subsystem
